[migration]: Migrate employment page
