+++ package.json
@@ -37,6 +37,7 @@
     "lodash": "^4.17.21",
     "luxon": "^3.4.4",
     "next": "^14.1.0",
+    "nuqs": "^1.17.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sharp": "^0.33.2",
