+++ src/components/employment/index.ts
@@ -1,5 +1,6 @@
 import EmploymentCard from "./EmploymentCard";
 import EmploymentDurationDisplay from "./EmploymentDurationDisplay";
 import EmploymentTypeSelect from "./EmploymentTypeSelect";
+export { EmploymentSelection } from "./EmploymentSelection";
 
 export { EmploymentCard, EmploymentDurationDisplay, EmploymentTypeSelect };
