+++ src/components/employment/styled/EmploymentBackground.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Box } from "@mui/material";
 import { styled } from "@mui/material/styles";
