+++ src/components/employment/styled/EmploymentCardContainer.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Card } from "@mui/material";
 import { styled } from "@mui/material/styles";
