+++ src/components/employment/styled/EmploymentCardMedia.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { CardMedia } from "@mui/material";
 import { styled } from "@mui/material/styles";
