+++ src/components/employment/styled/EmploymentJobChip.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Chip } from "@mui/material";
 import { styled } from "@mui/material/styles";
