+++ src/components/employment/styled/EmploymentPageHeaderContainer.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { alpha, Box } from "@mui/material";
 import { styled } from "@mui/material/styles";
