+++ src/components/employment/styled/EmploymentResumeLink.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { styled } from "@mui/material/styles";
 import Link from "next/link";
