+++ src/components/employment/styled/EmploymentText.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
